调整创建订单逻辑

+++ standalone-bis/src/main/java/com/eryi/bean/dto/OrderItemDto.java
@@ -27,7 +27,7 @@ public class OrderItemDto {
     /**
      * 数量
      */
-    private BigDecimal count;
+    private int count;
     /**
      * 运费
      */
